Finally got it done... fully functional version of node-osc compatible with node v0.6.13
