+++ lib/osc.js
@@ -1,19 +1,19 @@
-var buffer = require('buffer');
+var min = require('osc-min');
 var dgram = require('dgram');
 var util = require('util');
 var events = require('events');
 var jspack = require('jspack').jspack;
-var min = require('osc-min')
 
 ////////////////////
 // OSC Message
 ////////////////////
 
-function Message (address) {
-    this.address = address;
-    this.typetags = '';
-    this.args = [];
 
+function Message(address) {
+    this.oscType = "message";
+    this.address = address;
+    this.args = [];
+    
     for (var i = 1; i < arguments.length; i++) {
         this.append(arguments[i]);
     }
@@ -23,8 +23,7 @@ Message.prototype = {
     append: function (arg) {
         switch (typeof arg) {
         case 'object':
-            if (arg.typetag) {
-                this.typetags += arg.typetag;
+            if (arg.type) {
                 this.args.push(arg);
             } else {
                 throw new Error("don't know how to encode object " + arg)
@@ -32,101 +31,59 @@ Message.prototype = {
             break;
         case 'number':
             if (Math.floor(arg) == arg) {
-                this.typetags += TInt.prototype.typetag;
-                this.args.push(new TInt(Math.floor(arg)));
+                var argOut = new Argument('integer', arg);
+                this.args.push(argOut);
             } else {
-                this.typetags += TFloat.prototype.typetag;
-                this.args.push(new TFloat(arg));
+                var argOut = new Argument('float', arg);
+                this.args.push(argOut);
             }
             break;
         case 'string':
-            this.typetags += TString.prototype.typetag;
-            this.args.push(new TString(arg));
+            var argOut = new Argument('string', arg);
+            this.args.push(argOut);
             break;
         default:
             throw new Error("don't know how to encode " + arg);
-        }
-    },
-
-    toBinary: function () {
-        var address = new TString(this.address);
-        var binary = [];
-        var pos = 0;
-        pos = address.encode(binary, pos);
-        if (this.typetags) {
-            var typetags = new TString(',' + this.typetags);
-            pos = typetags.encode(binary, pos);
-            for (var i = 0; i < this.args.length; i++) {
-                pos = this.args[i].encode(binary, pos);
-            }
-        }
-        return binary;
-    },
+        }   
+    }
 }
+
 exports.Message = Message;
 
-var Bundle = function (time) {
-    this.timetag = time || 0;
-    this.parts = [];
-    for (var i = 1; i < arguments.length; i++) {
-        this.append(arguments[i]);
-    }
+function Argument(type, value){
+    this.type = type;
+    this.value = value;
 }
-
-util.inherits(Bundle, Message);
-
-Bundle.prototype.append = function (arg) {
-    var message;
-    if (arg instanceof Message) {
-        message = arg;
-    } else {
-        // cheesy
-        var message = new Message;
-        Message.apply(message, arguments);
-    }
-    this.parts.push(message);
-}
-
-Bundle.prototype.toBinary = function () {
-    var buffer = [];
-    var p = 0;
-    p = (new TString('#bundle')).encode(buffer, p);
-    p = (new TTime(this.timetag)).encode(buffer, p);
-    for (var i = 0; i < this.parts.length; i++) {
-        var binary = this.parts[i].toBinary();
-        var length = binary.length;
-        jspack.PackTo('>i', buffer, buffer.length, [length]);
-        buffer = buffer.concat(binary);
-    }
-    return buffer;
-}
-
-exports.Bundle = Bundle;
 
 ////////////////////
 // OSC Client
 ////////////////////
 
 var Client = function (host, port) {
-    this.port = port;
     this.host = host;
+    this.port = port;
     this._sock = dgram.createSocket('udp4');
 }
 
 Client.prototype = {
-    send: function () {
-        var binary;
-        if (arguments[0].toBinary) {
-            binary = arguments[0].toBinary();
-        } else {
-            // cheesy
-            var message = new Message;
-            Message.apply(message, arguments)
-            binary = Message.prototype.toBinary.call(message);
+    send: function (message) {
+        switch (typeof message) {
+            case 'object':
+                var buf = min.toBuffer(message);
+                this._sock.send(buf, 0, buf.length, this.port, this.host);
+                break;
+            case 'string':
+                mes = new Message(arguments[0]);
+                for (var i = 1; i < arguments.length; i++) {
+                    mes.append(arguments[i]);
+                }
+                var buf = min.toBuffer(mes);
+                this._sock.send(buf, 0, buf.length, this.port, this.host);
+                break;
+            default:
+                throw new Error("That Message Just Doesn't Seem Right");
         }
-        var b = new buffer.Buffer(binary, 'binary');
-        this._sock.send(b, 0, b.length, this.port, this.host);
-    }                                                       // 
+    }                                                  
 }
 
 exports.Client = Client;
@@ -319,6 +276,7 @@ function decode (data) {
 };
 
 
+
 ////////////////////
 // OSC Server
 ////////////////////
@@ -349,3 +307,5 @@ var Server = function(port, host) {
 util.inherits(Server, events.EventEmitter);
 
 exports.Server = Server;
+
+
