+++ server.js
@@ -1,7 +1,8 @@
 var osc = require('./lib/osc');
 
+
 var oscServer = new osc.Server(3333, '0.0.0.0');
 oscServer.on("message", function (msg, rinfo) {
-	console.log("TUIO message:");
+	console.log("Message:");
 	console.log(msg);
 });
